Moved Recolorable to Launcher
