+++ Launcher/Recolorable.java
@@ -1,4 +1,4 @@
-package GUI;
+package Launcher;
 
 public interface Recolorable {
     void resetColor();
